refactor: Move orientSolver helper functions to own module
